# api/openapi.yaml
openapi: 3.0.3
info:
  title: RefMD API
  description: Real-time collaborative markdown editor API with CRDT support
  version: 1.1.0
  contact:
    name: RefMD Team
    email: support@refmd.io
servers:
  - url: http://localhost:8888/api
    description: Development server
  - url: https://api.refmd.io/api
    description: Production server
paths:
...
  /shares/documents/{id}/share:
    post:
      tags:
        - Sharing
      summary: Create share link
      operationId: shareDocument
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ShareDocumentRequest'
      responses:
        '201':
          description: Share created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ShareResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /shares/documents/{id}/shares:
    get:
      tags:
        - Sharing
      summary: List document shares
      operationId: listDocumentShares
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Shares retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Share'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /shares/{token}:
    get:
      tags:
        - Sharing
      summary: Get shared document
      operationId: getSharedDocument
      parameters:
        - name: token
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Shared document retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SharedDocument'
        '404':
          $ref: '#/components/responses/NotFound'

    delete:
      tags:
        - Sharing
      summary: Delete share
      operationId: deleteShare
      security:
        - bearerAuth: []
      parameters:
        - name: token
          in: path
          required: true
          schema:
            type: string
      responses:
        '204':
          description: Share deleted successfully
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'

  # ===== Git Sync =====
...
  /scraps/{id}/share:
    post:
      tags:
        - Scraps
      summary: Create share link for scrap
      operationId: createScrapShare
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ShareDocumentRequest'
      responses:
        '201':
          description: Share link created
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/ShareResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
...
  /scraps/{id}/shares:
    get:
      tags:
        - Scraps
      summary: List share links for scrap
      operationId: listScrapShares
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: List of share links
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      type: object
                      properties:
                        id:
                          type: string
                          format: uuid
                        token:
                          type: string
                        document_id:
                          type: string
                          format: uuid
                        permission_level:
                          type: string
                        created_by:
                          type: string
                          format: uuid
                        expires_at:
                          type: string
                          format: date-time
                          nullable: true
                        created_at:
                          type: string
                          format: date-time
                        url:
                          type: string
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /scraps/shares/{token}:
    delete:
      tags:
        - Scraps
      summary: Delete scrap share link
      operationId: deleteScrapShare
      security:
        - bearerAuth: []
      parameters:
        - name: token
          in: path
          required: true
          schema:
            type: string
      responses:
        '204':
          description: Share link deleted
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'

  # ===== Public Documents =====
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ShareDocumentRequest:
      type: object
      properties:
        permission:
          type: string
          enum: [view, edit]
        expires_at:
          type: string
          format: date-time
          nullable: true
      required:
        - permission
    ShareResponse:
      type: object
      properties:
        data:
          type: object
          properties:
            token:
              type: string
            url:
              type: string
            expires_at:
              type: string
              format: date-time
              nullable: true
    Share:
      type: object
      properties:
        id:
          type: string
          format: uuid
        document_id:
          type: string
          format: uuid
        token:
          type: string
        permission:
          type: string
          enum: [view, edit]
        created_by:
          type: string
          format: uuid
        expires_at:
          type: string
          format: date-time
          nullable: true
        created_at:
          type: string
          format: date-time
    SharedDocument:
      type: object
      properties:
        id:
          type: string
          format: uuid
        title:
          type: string
        content:
          type: string
        permission:
          type: string
          enum: [view, edit]

    # ===== Git Sync =====
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: object
          properties:
            code:
              type: string
            message:
              type: string
            details:
              type: array
              items:
                type: object
...
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Forbidden:
      description: Forbidden
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    NotFound:
      description: Not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...